「新規作成」と「編集」のviewとcontrollerを統一
