+++ RemeApp/Controller/EditShoppingListView/EditItemViewController.swift
@@ -1,17 +1,19 @@
 //
-//  EditItemViewController.swift
+//  CreateNewItemViewController.swift
 //  RemeApp
 //
-//  Created by 本川晴彦 on 2023/04/09.
+//  Created by 本川晴彦 on 2023/04/05.
 //
 
 import UIKit
 import RealmSwift
-
-/// I-品目編集
+/// G-品目新規作成
 class EditItemViewController: UIViewController {
 
     // MARK: - @IBOutlet & @IBAction
+
+    /// タイトル
+    @IBOutlet weak var titleLabel: UILabel!
     /// 商品名入力
     @IBOutlet private weak var nameOfItemTextField: UITextField!
     /// 個数入力
@@ -37,10 +39,40 @@ class EditItemViewController: UIViewController {
     @IBOutlet private weak var supplementTextView: UITextView!
     /// 写真選択ボタン
     @IBOutlet private weak var selectPhotoButton: UIButton!
-    /// 写真を追加するアクションを実行
+    /// カメラ撮影とフォトライブラリーでの写真選択を実行する処理
+    /// - アクションシートで選択
+    /// - カメラ撮影アクション
+    /// - フォトライブリーラリーから選択アクション
+    /// - キャンセルアクション
     @IBAction private func addPhoto(_ sender: Any) {
-        // カメラまたはフォトライブラリー、キャンセルを選択するメソッド
-        setCameraAndPhotoAction()
+        /// アラートコントローラーをインスタンス化
+        let alertController = UIAlertController(title: "選択して下さい", message: nil, preferredStyle: .actionSheet)
+        // カメラ撮影アクション定義
+        if UIImagePickerController.isSourceTypeAvailable(.camera) {
+            let cameraAction = UIAlertAction(title: "カメラ", style: .default, handler: { (action) in
+                let imagePickerController = UIImagePickerController()
+                imagePickerController.sourceType = .camera
+                imagePickerController.delegate = self
+                self.present(imagePickerController, animated: true, completion: nil)
+            })
+            alertController.addAction(cameraAction)
+        }
+        // フォトラーブラリーから選択アクションを定義
+        if UIImagePickerController.isSourceTypeAvailable(.photoLibrary) {
+            let photoLibraryAction = UIAlertAction(title: "フォトライブラリー", style: .default,handler: { (action) in
+                let imagePickerController = UIImagePickerController()
+                imagePickerController.sourceType = .photoLibrary
+                imagePickerController.delegate = self
+                self.present(imagePickerController, animated: true)
+            })
+            alertController.addAction(photoLibraryAction)
+        }
+        // キャンセルアクションを定義
+        let cancelAction = UIAlertAction(title: "キャンセル", style: .cancel, handler: nil)
+        alertController.addAction(cancelAction)
+        // iPadでの処理落ち防止処置
+        alertController.popoverPresentationController?.sourceView = view
+        present(alertController, animated: true)
     }
     /// 写真削除ボタン
     @IBOutlet private weak var deletePhotoButton: UIButton!
@@ -72,17 +104,25 @@ class EditItemViewController: UIViewController {
     private let unitArray: Array<String> = ["個", "本", "袋", "パック"]
     /// 写真のURLパス
     private var imageFilePath: URL?
+    /// カスタム売り場マップのリスト
+    private var customSalesFloorData = CustomSalesFloorModel()
+
+    /// お使いデータ
+    var errandData = ErrandDataModel()
+    // デリゲート
+    var delegate: EditItemViewControllerDelegate?
+
     /// nameOfItemTextFieldに表示するテキスト
-    private var nameOfItemTextFieldText:String = ""
+    private var nameOfItemTextFieldText:String? = nil
 
     /// numberOfItemPickerViewに表示する文字列
-    private var numberOfItemPickerViewString:String = ""
+    private var numberOfItemPickerViewString:String = "１"
 
     /// unitPickerViewに表示する文字列
-    private var unitPickerViewString:String = ""
+    private var unitPickerViewString:String = "個"
 
-    /// selectTypeOfSalesFloorButtonに表示する売り場の種類を指定するためのRawValue
-    private var selectedSalesFloorRawValue:Int = 0
+    ///  売り場を保存するための一時置き場
+    private var selectedSalesFloorRawValue:Int? = nil
 
     /// supplementTextViewに表示するテキスト
     private var supplementTextViewText:String? = nil
@@ -90,26 +130,20 @@ class EditItemViewController: UIViewController {
     /// photoImageViewに表示する画像
     private var photoImageViewImage:UIImage? = nil
 
-    /// お使いデータ
-    var errandData = ErrandDataModel()
-
-    /// カスタム売り場マップのリスト
-    private var customSalesFloorData = CustomSalesFloorModel()
-
     // MARK: - viewDidLoad
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.navigationItem.title = "品目編集"
         setDataSourceAndDelegate()
         setKeyboardCloseButton()
         setAppearanceAllButton()
+        displayData()
+        setTitleLabel()
         setDisableOrEnable()
         supplementTextView.setAppearance()
-        displayData()
     }
 
     // MARK: - func
-    /// データソースとデリゲート設定
+    /// データソースとデリゲートをセット
     private func setDataSourceAndDelegate() {
         numberOfItemPickerView.delegate = self
         numberOfItemPickerView.dataSource = self
@@ -119,16 +153,6 @@ class EditItemViewController: UIViewController {
         supplementTextView.delegate = self
     }
 
-    /// 画面遷移してきた際に写真データの有無で写真の削除ボタンの活性化、非活性化を切り替える処理
-    private func setDisableOrEnable() {
-        // もし写真がない場合
-        if photoImageViewImage == nil {
-            deletePhotoButton.setDisable()
-        } else {
-            deletePhotoButton.setEnable()
-        }
-    }
-
     /// 画面上の全てのButtonの見た目の設定メソッド
     private func setAppearanceAllButton() {
         selectTypeOfSalesFloorButton.setAppearanceWithShadow()
@@ -136,6 +160,30 @@ class EditItemViewController: UIViewController {
         cancelButton.setAppearanceWithShadow()
         addButton.setAppearanceWithShadow()
         deletePhotoButton.setAppearanceWithShadow()
+    }
+
+    private func setTitleLabel() {
+        if nameOfItemTextField.text == "" {
+            titleLabel.text = "新規作成"
+        } else {
+            titleLabel.text = "編集"
+        }
+    }
+
+    /// 画面遷移してきた際に写真データの有無で写真の削除ボタンの活性化、非活性化を切り替える処理
+    private func setDisableOrEnable() {
+        if nameOfItemTextField.text == "" {
+            selectTypeOfSalesFloorButton.setDisable()
+            addButton.setDisable()
+        } else {
+            addButton.setEnable()
+        }
+        // もし写真がない場合
+        if photoImageView.image == nil {
+            deletePhotoButton.setDisable()
+        } else {
+            deletePhotoButton.setEnable()
+        }
     }
 
     /// データ受け渡し用のメソッド
@@ -159,22 +207,11 @@ class EditItemViewController: UIViewController {
         photoImageView.image = photoImageViewImage
     }
 
-    /// numberOfItemPickerViewに表示できるように変換する
-    func selectNumberOfItemRow(selectedNumberOfItem: String) {
-        let numberOfItemIndex = numberOfItemArray.firstIndex(of: selectedNumberOfItem) ?? 0
-        numberOfItemPickerView.selectRow(numberOfItemIndex, inComponent: 0, animated: true)
-    }
-
-    /// unitPickerViewに表示できるように変換する
-    func selectUnitRow(selectedUnit: String) {
-        let selectedUnitIndex = unitArray.firstIndex(of: selectedUnit) ?? 0
-        unitPickerView.selectRow(selectedUnitIndex, inComponent: 0, animated: true)
-    }
-
     /// salesFloorTypeButtonに売り場の内容を反映させる
     /// - 売り場の名称を設定
     /// - 売り場の色を設定
-    func setSalesFloorTypeButton(salesFloorRawValue: Int) {
+    func setSalesFloorTypeButton(salesFloorRawValue: Int?) {
+        guard let salesFloorRawValue else { return }
         let useSalesFloorTypeKey = "useSalesFloorTypeKey"
         let salesFloorTypeInt = UserDefaults.standard.integer(forKey: useSalesFloorTypeKey)
         // 0 -> カスタム、1(else) -> デフォルト
@@ -233,6 +270,18 @@ class EditItemViewController: UIViewController {
         }
     }
 
+    /// numberOfItemPickerViewに表示できるように変換する
+    func selectNumberOfItemRow(selectedNumberOfItem: String) {
+        let numberOfItemIndex = numberOfItemArray.firstIndex(of: selectedNumberOfItem) ?? 0
+        numberOfItemPickerView.selectRow(numberOfItemIndex, inComponent: 0, animated: true)
+    }
+
+    /// unitPickerViewに表示できるように変換する
+    func selectUnitRow(selectedUnit: String) {
+        let selectedUnitIndex = unitArray.firstIndex(of: selectedUnit) ?? 0
+        unitPickerView.selectRow(selectedUnitIndex, inComponent: 0, animated: true)
+    }
+
     /// キーボードの完了ボタン配置、完了ボタン押してキーボードを非表示に変更するメソッド
     private func setKeyboardCloseButton() {
         let toolbar = UIToolbar(frame: CGRect(x: 0, y: 0, width: 320, height: 40))
@@ -241,7 +290,6 @@ class EditItemViewController: UIViewController {
         nameOfItemTextField.inputAccessoryView = toolbar
         supplementTextView.inputAccessoryView = toolbar
     }
-
     /// 閉じるボタンを押した時にキーボードを閉じるメソッド
     @objc func doneButtonTapped() {
         view.endEditing(true)
@@ -253,12 +301,12 @@ extension EditItemViewController {
     /// アラートで確認するメソッド
     /// - 編集を中止て前の画面に戻るか
     /// - 中止をキャンセルして画面に止まるか
-    func showCancelAlert() {
+    private func showCancelAlert() {
         let alertController = UIAlertController(title: "編集を中止", message:
                                                     "編集内容を破棄してもよろしいですか？", preferredStyle: .alert)
         // はいでEditShoppingListViewに戻る
         let okAction = UIAlertAction(title: "はい", style: .default, handler:  { (action) in
-            self.navigationController?.popViewController(animated: true)
+            self.dismiss(animated: true)
         })
         // キャンセル、何もしない
         let cancelAction = UIAlertAction(title: "キャンセル", style: .cancel, handler: nil)
@@ -271,7 +319,7 @@ extension EditItemViewController {
     /// 追加ボタンをタップした時の処理
     /// - 商品名が未入力の場合はアラートを出す
     /// - 商品名が入力されていればデータを書き込み、画面を閉じる
-    func addOrReEnter() {
+    private func addOrReEnter() {
         if nameOfItemTextField.text == "" {
             // 警告アラート
             let alertController = UIAlertController(title: nil, message:
@@ -283,7 +331,8 @@ extension EditItemViewController {
         } else {
             // ここに追加の処理
             saveData()
-            self.navigationController?.popViewController(animated: true)
+            self.dismiss(animated: true)
+            delegate?.savedReload()
         }
     }
 
@@ -292,13 +341,13 @@ extension EditItemViewController {
         let selectedNumberOfItem = numberOfItemArray[numberOfItemPickerView.selectedRow(inComponent: 0)]
         // numberOfItemPickerViewで選択された値を取得
         let selectedUnit = unitArray[unitPickerView.selectedRow(inComponent: 0)]
-
+        // データベースに保存
         let realm = try! Realm()
         try! realm.write {
             errandData.nameOfItem = nameOfItemTextField.text!
             errandData.numberOfItem = selectedNumberOfItem
             errandData.unit = selectedUnit
-            errandData.salesFloorRawValue = selectedSalesFloorRawValue
+            errandData.salesFloorRawValue = selectedSalesFloorRawValue!
             if supplementTextView.text == "" {
                 errandData.supplement = nil
             } else {
@@ -311,8 +360,8 @@ extension EditItemViewController {
     }
 }
 
-// MARK: - UIPickerViewDelegate&DataSource
-extension EditItemViewController: UIPickerViewDelegate, UIPickerViewDataSource {
+// MARK: - UIPickerViewDataSource&Delegate
+extension EditItemViewController: UIPickerViewDataSource, UIPickerViewDelegate {
     /// pickerViewに表示する内容
     func pickerView(_ pickerView: UIPickerView, titleForRow row: Int, forComponent component: Int)
     -> String? {
@@ -325,7 +374,6 @@ extension EditItemViewController: UIPickerViewDelegate, UIPickerViewDataSource {
                 return nil
         }
     }
-
     /// pickerViewの列の数
     func numberOfComponents(in pickerView: UIPickerView) -> Int {
         return 1
@@ -342,7 +390,6 @@ extension EditItemViewController: UIPickerViewDelegate, UIPickerViewDataSource {
         }
     }
 }
-
 
 // MARK: - UITextFieldDelegate
 extension EditItemViewController: UITextFieldDelegate {
@@ -367,9 +414,21 @@ extension EditItemViewController:SelectTypeOfSalesFloorViewControllerDelegate {
     /// SelectTypeOfSalesFloorViewで各Buttonをタップした際のメソッド
     /// - selectTypeOfSalesFloorButtonのタイトルを該当する売り場の名称に変更
     /// - selectTypeOfSalesFloorButtonのバックグラウンドカラーを該当する売り場の色に変更
+    /// - addButtonを活性化
     func salesFloorButtonDidTapDone(salesFloorRawValue: DefaultSalesFloorType.RawValue) {
+        let useSalesFloorTypeKey = "useSalesFloorTypeKey"
+        let salesFloorTypeInt = UserDefaults.standard.integer(forKey: useSalesFloorTypeKey)
+        // 0 -> カスタム、1(else) -> デフォルト
+        if salesFloorTypeInt == 0 {
+            // カスタムマップタイプの処理
+            setCustomSalesFloorButton(salesFloorRawValue: salesFloorRawValue)
+            addButton.setEnable()
+        } else {
+            // デフォルトマップタイプの処理
+            setDefaultSalesFloorButton(salesFloorRawValue: salesFloorRawValue)
+            addButton.setEnable()
+        }
         selectedSalesFloorRawValue = salesFloorRawValue
-        setSalesFloorTypeButton(salesFloorRawValue: salesFloorRawValue)
     }
 }
 
@@ -377,9 +436,9 @@ extension EditItemViewController:SelectTypeOfSalesFloorViewControllerDelegate {
 extension EditItemViewController: UITextViewDelegate {
     /// 入力があったらプレースホルダーのラベルを非表示
     func textViewDidBeginEditing(_ textView: UITextView) {
-        placeholderLabel.isHidden = true
+            placeholderLabel.isHidden = true
     }
-    /// 何も入力されていなかったらプレースホルダーをセット
+    /// 編集終了後、何も入力されていなかったらプレースホルダーをセット
     func textViewDidEndEditing(_ textView: UITextView) {
         if supplementTextView.text == "" {
             placeholderLabel.isHidden = false
@@ -388,7 +447,8 @@ extension EditItemViewController: UITextViewDelegate {
         }
     }
     /// 入力制限を３０文字以内で設定
-    func textView(_ textView: UITextView, shouldChangeTextIn range: NSRange, replacementText text: String) -> Bool {
+    func textView(_ textView: UITextView, shouldChangeTextIn range: NSRange,
+                  replacementText text: String) -> Bool {
         let maxLength = 30
         let currentString: NSString = supplementTextView.text as NSString
         let updatedString = currentString.replacingCharacters(in: range, with: text)
@@ -396,13 +456,13 @@ extension EditItemViewController: UITextViewDelegate {
     }
 }
 
-// MARK: - UIImagePickerControllerDelegate&UINavigationControllerDelegate
+// MARK: - UIImagePickerDelegate,UINavigationControllerDelegate
 // !!!: いずれPHPickerに変更しないといけないかも
 // 写真添付と削除の処理関連
 extension EditItemViewController: UIImagePickerControllerDelegate, UINavigationControllerDelegate {
     /// 撮影または写真選択が終了した際のメソッド
     /// - 撮影または選択した画像をUIImageViewに表示させる
-    /// - 写真の取り消しボタンを有効化する
+    /// - 写真の取り消しボタンを活性化する
     /// - 画面を閉じてCreateNewItemViewに戻る
     func imagePickerController(_ picker: UIImagePickerController,
                                didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
@@ -437,5 +497,7 @@ extension EditItemViewController: UIImagePickerControllerDelegate, UINavigationC
     }
 }
 
-// MARK: - CameraAndPhotoActionable
-extension EditItemViewController: CameraAndPhotoActionable {}
+// MARK: - EditItemViewControllerDelegate
+protocol EditItemViewControllerDelegate {
+    func savedReload()
+}
